requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
requests==2.32.3
matplotlib==3.8.4
